+++ institutemanagementsystem/src/main/java/com/winhe/institute/management/lecturerattendance/LecturerAttendanceService.java
@@ -3,6 +3,7 @@ package com.winhe.institute.management.lecturerattendance;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.winhe.institute.management.associate.student_batch.Student_Batch;
 import com.winhe.institute.management.util.jsonwrapper.JsonWrapper;
 
 @Service
@@ -32,24 +33,48 @@ public class LecturerAttendanceService {
 
 	}
 
-	// method of updating a lecturer attendance
-	public JsonWrapper updateLecturerAttendance(LecturerAttendance lecturerAttendance) {
+	// method of adding a lecturer attendance
+	public JsonWrapper addLecturerAttendance(LecturerAttendance lecturerAttendance) {
 
 		lecturerAttendanceRepository.save(lecturerAttendance);
 
 		JsonWrapper data = new JsonWrapper("LECTURERATTENDANCE200", "SUCCESSFULLY ADDED", lecturerAttendance);
 
 		return data;
-
 	}
+
+	// method of updating a lecturer attendance
+	public JsonWrapper updateLecturerAttendance(Long id, LecturerAttendance lecturerAttendance) {
+
+		LecturerAttendance tempLecturerAttendance = lecturerAttendanceRepository.findOne(id);
+
+		if (tempLecturerAttendance == null) {
+			JsonWrapper data = new JsonWrapper("LECTURERATTENDANCE404", "ERROR , No Lecturer Attendance Entry Found");
+			return data;
+		}
+
+		lecturerAttendance.setId(id);
+		lecturerAttendanceRepository.save(lecturerAttendance);
+		
+		JsonWrapper data = new JsonWrapper("LECTURERATTENDANCE200", "Lecturer Attendance Entry Successfully Updated", lecturerAttendance);
+
+		return data;
+	};
 
 	public JsonWrapper deleteLecturerAttendance(Long id) {
 
+		LecturerAttendance templecturerAttendance= lecturerAttendanceRepository.findOne(id);
+
+		if (templecturerAttendance == null) {
+			JsonWrapper data = new JsonWrapper("LECTURERATTENDANCE404", "ERROR , No Lecturer Attendance Entry Found");
+			return data;
+		}
+
 		lecturerAttendanceRepository.delete(id);
 
-		JsonWrapper data = new JsonWrapper("LECTURERATTENDANCE200", "SUCCESSFULLY ADDED" + id);
+		JsonWrapper data = new JsonWrapper("LECTURERATTENDANCE200", "SUCCESFULLY DELETED");
 
 		return data;
-	}
+	};
 
 }
